% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sae_ml_linear.R
\name{as.data.frame.sae_ml_linear}
\alias{as.data.frame.sae_ml_linear}
\title{Coerce an sae_ml_linear object to a data frame}
\usage{
\method{as.data.frame}{sae_ml_linear}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{Object of class \code{"sae_ml_linear"}.}

\item{row.names}{Ignored.}

\item{optional}{Ignored.}

\item{...}{Further arguments.}
}
\value{
A data frame containing domain-level estimates.
}
\description{
Coerce an sae_ml_linear object to a data frame
}
