% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sae_ml_linear.R
\name{print.sae_ml_linear}
\alias{print.sae_ml_linear}
\title{Print method for sae_ml_linear}
\usage{
\method{print}{sae_ml_linear}(x, n = 6L, ...)
}
\arguments{
\item{x}{Object of class \code{"sae_ml_linear"}.}

\item{n}{Number of rows to display.}

\item{...}{Further arguments.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for sae_ml_linear
}
