% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sae_ml_linear.R
\name{summary.sae_ml_linear}
\alias{summary.sae_ml_linear}
\title{Summary method for sae_ml_linear}
\usage{
\method{summary}{sae_ml_linear}(object, n = 6L, ...)
}
\arguments{
\item{object}{Object of class \code{"sae_ml_linear"}.}

\item{n}{Number of rows to display.}

\item{...}{Further arguments.}
}
\value{
Invisibly returns \code{object}.
}
\description{
Summary method for sae_ml_linear
}
