% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sae_ml_linear.R
\name{print.sae_ml_linear}
\alias{print.sae_ml_linear}
\title{Print method for sae_ml_linear}
\usage{
\method{print}{sae_ml_linear}(x, n = 6L, ...)
}
\arguments{
\item{x}{An object of class \code{"sae_ml_linear"}.}

\item{n}{Integer. Number of rows to preview from \code{estimates}.}

\item{...}{Further arguments (currently unused).}
}
\description{
Displays a concise summary: title, formula, estimator, number of domains,
and a preview of the final estimates table.
}
