% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sae_ml_linear.R
\name{summary.sae_ml_linear}
\alias{summary.sae_ml_linear}
\title{Summary method for sae_ml_linear}
\usage{
\method{summary}{sae_ml_linear}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"sae_ml_linear"}.}

\item{...}{Further arguments (currently unused).}
}
\description{
Summary method for sae_ml_linear
}
